midi_uart/midi_uart/requirements.txt
pyserial
mido
